pyproject.toml
[tool.poetry]
name = "telegram_mailing_help"
version = "0.1.0"
description = ""
authors = ["cherkashin <alexander.cherkashin@sibit.tech>"]

[tool.poetry.dependencies]
python = "^3.8"
python-telegram-bot = "^13.1"
bottle = "^0.12.19"
yoyo-migrations = "^7.2.1"
dacite = "^1.6.0"
sqlite3worker = "^1.1.7"
gevent = "^20.12.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
